MI-63 Architecture modification: mappers style is used to commnicate with the server
