+++ UnitTests/UnitTests.csproj
@@ -55,7 +55,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="MainActivity.cs" />
-    <Compile Include="MRidesWebApiTest.cs" />
+    <Compile Include="MappersTest\ConsoleMapperTests.cs" />
+    <Compile Include="MappersTest\RequestMapperTests.cs" />
+    <Compile Include="MappersTest\UserMapperTests.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
@@ -79,6 +81,7 @@
       <Name>mRides-app</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
      Other similar extension points exist, see Microsoft.Common.targets.
