+++ mRides-app/MainActivity.cs
@@ -81,7 +81,10 @@ namespace mRides_app
                     //userProfileActivity.PutExtra("Profile Info", userName);
                     //StartActivity(userProfileActivity);
 
-                    /** GO TO PREFERENCES */
+                    /** 
+                     * If we already know the user, instantiate the user object and skip the preference activity 
+                     * Otherwise, give the facebook ID to the preference activity (first time user)
+                     */
                     userName = obj["name"].ToString();
                     var preferencesActivity = new Intent(this, typeof(PreferencesActivity));
                     preferencesActivity.PutExtra(GetString(Resource.String.ExtraData_FacebookId), obj["id"].ToString());
