+++ mRides-app/mRides-app.csproj
@@ -112,21 +112,22 @@
     <Compile Include="Models\Ride.cs" />
     <Compile Include="Models\RiderRequest.cs" />
     <Compile Include="Models\UserRides.cs" />
-    <Compile Include="MRidesJSON\FindRidersJson.cs" />
-    <Compile Include="MRidesWebApi.cs" />
+    <Compile Include="Mappers\MRidesJSON\FindDriversJson.cs" />
+    <Compile Include="Mappers\MRidesJSON\FindRidersJson.cs" />
+    <Compile Include="Mappers\AbstractMapper.cs" />
+    <Compile Include="Mappers\ConsoleMapper.cs" />
+    <Compile Include="Mappers\RequestMapper.cs" />
+    <Compile Include="Mappers\UserMapper.cs" />
     <Compile Include="PreferencesActivity.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestMe.cs" />
     <Compile Include="UserTypeFragment.cs" />
     <Compile Include="Models\User.cs" />
-    <Compile Include="User.cs" />
     <Compile Include="UserProfile.cs" />
     <Compile Include="Feedback\FeedbackTest.cs" />
     <Compile Include="Feedback\DriverReview.cs" />
     <Compile Include="Feedback\RiderReview.cs" />
     <Compile Include="TestFragments.cs" />
-    <Compile Include="UserTypeFragment.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="GettingStarted.Xamarin" />
